requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
boto==2.49.0
boto3==1.9.165
botocore==1.12.165
certifi==2019.3.9
chardet==2.3.0
construct==2.5.3
cssselect==1.0.3
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-ckeditor==5.7.1
django-haystack==2.8.1
django-js-asset==1.2.2
django-json-ld==0.0.4
django-meta==1.5.1
django-taggit==1.1.0
docutils==0.14
docx2txt==0.6
EbookLib==0.15
elasticsearch==1.7.0
future==0.17.1
gensim==3.7.3
geoip2==2.9.0
idna==2.8
jmespath==0.9.4
lxml==4.3.3
maxminddb==1.4.1
nltk==3.4.3
numpy==1.16.4
oauthlib==3.0.1
pdf2image==1.5.4
pefile==2019.4.18
Pillow==6.0.0
pocketsphinx==0.1.3
pyquery==1.4.0
python-dateutil==2.8.0
python-pptx==0.6.5
python-ptrace==0.9.3
python3-openid==3.1.0
pytz==2019.1
rake-nltk==1.0.4
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
scipy==1.3.0
selenium==3.141.0
six==1.10.0
smart-open==1.8.4
SpeechRecognition==3.6.3
sqlparse==0.3.0
textract==1.6.1
urllib3==1.25.3
xlrd==1.0.0
XlsxWriter==1.1.8
